fixed pen-tracking bug
I was sending draw events in a requestAnimFrame function, which was
async, so what would happen is locally, there would be immediate pen-up
event, then a fraction of a second later, if this happened between
frames, we'd send another draw event, which basically would put a dot
and the client(s) would think we had our pen down. this would cause the
next event to come through to connect to the last point, which is
annoying.