fixed part of a bug in iOS with drawing
apparently ios can't drawImage outside of the bounds of the source
image... so if the sx or sy is negative, it just fails to draw entirely.

this code makes sure that the x/y drawing is fixed. if you scroll off
the image on the right or bottom, it still fails, so that's still here
to fix.