properly identify users who are disconnecting
when counting up how many times a user is connected, base it on the
user_id, not the pid. when a user is connected more than once to a
board, and disconnects one of the clients, it will no longer broadcast
that they left even though they are still connected.